refactored of css and import component in index.js
