--- index.css
@@ -1,11 +1,0 @@
-:root{
-    --primary-color: rgb(173, 171, 171);
-    --default-color: white;
-    --loginbackground-color: rgb(219, 219, 219);
-    --signbackground-color: rgb(243, 243, 243);
-    --footerbackground-color: rgb(53, 52, 52);
-    --buybutton-color: rgb(116, 114, 114);
-    --discount-color: rgb(26, 192, 26);
-    --wislist-color: red;
-    --card-border: rgb(233, 231, 231);
-}
