+++ src/App.js
@@ -1,22 +1,19 @@
+import { Footer, Navbar } from 'component'
+import { Homepage } from 'pages'
 import React from 'react'
-import { Footer } from './component/footer/Footer'
-import { Navbar } from './component/navbar/Navbar'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import { Homepage } from 'pages/homepage/Homepage'
+
 
 export const App = () => {
   return (
-    
-  
-   
     <>
-    <BrowserRouter>
-    <Navbar />
-      <Routes>
-        <Route path='/' element={<Homepage />} />
-      </Routes>
-    <Footer />
-    </BrowserRouter>
-  </>
+      <BrowserRouter>
+          <Navbar />
+            <Routes>
+              <Route path='/' element={<Homepage />} />
+            </Routes>
+          <Footer />
+      </BrowserRouter>
+    </>
   )
 }
