+++ src/component/footer/Footer.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import './footer.css'
-// import { Link } from 'react-router-dom'
 
 export const Footer = () => {
     return (
