+++ src/component/footer/footer.css
@@ -8,7 +8,7 @@
     width: 100%;
     text-align: center;
     line-height: 1.5rem;
-    font-size:  1rem;
+    font-size:  1.3rem;
     box-shadow: 0px 0px 5px rgb(139, 139, 139);
     padding: 5px 0 5px 0;
