+++ src/component/index.js
@@ -1,0 +1,3 @@
+export { Navbar } from "./navbar/Navbar";
+export { Footer } from "./footer/Footer";
+export { QuizCard } from "./card/QuizCard";
