+++ src/pages/homepage/Homepage.js
@@ -1,4 +1,4 @@
-import { QuizCard } from 'component/card/QuizCard'
+import { QuizCard } from 'component'
 import React from 'react'
 import '../../component/card/quizcard.css'
 
@@ -6,8 +6,8 @@ import '../../component/card/quizcard.css'
 export const Homepage = () => {
   return (
     <div className="category-card">
-    <QuizCard 
-    heading="Mathmatics"
+
+    <QuizCard heading="Mathmatics"
               about= "Challenge yourself to take this quiz on Indian History??? I bet you cannot get more than an average child."   />
     <QuizCard heading="Mathmatics"
               about= "Challenge yourself to take this quiz on Indian History??? I bet you cannot get more than an average child."   />
